+++ tests/testthat.R
@@ -1,0 +1,5 @@
+library(testthat)
+library(HIMA)
+
+test_check("HIMA")
+